refactor(sync-core): split CrossTabSocket into focused modules
The single-file implementation has grown to ~1k lines doing five
different jobs. Now that the feature has settled, split it into modules
with sharp boundaries — each piece is independently testable and
replaceable.

  packages/sync-core/src/lib/cross-tab/
  ├── protocol.ts        Message envelope + injectable interfaces
  ├── browser-env.ts     CrossTabBrowserEnv + default impl
  ├── Leader.ts          Web Lock + visibility-driven want-lock
  ├── Presenter.ts       Focus-driven claim + $isPresenter signal
  ├── LeaderRouter.ts    clientClock remap + push_result synthesis
  └── CrossTabSocket.ts  Glue: implements TLPersistentClientSocket

Also moves the whole thing from `sync` to `sync-core` — the right home
alongside ClientWebSocketAdapter and the other TLPersistentClientSocket
implementations. `useSync` imports CrossTabSocket from sync-core now.

Each module is a small enough surface that an alternate strategy (a
SharedWorker-based Leader, a different Presenter heuristic, a v2-
protocol LeaderRouter) can drop in without touching the others.

Pure restructure — no behavior change, no test changes besides import
paths.
