+++ packages/sync-core/src/lib/cross-tab/protocol.ts
@@ -1,0 +1,113 @@
+import { TLRecord } from '@tldraw/tlschema'
+import { TLSocketClientSentEvent, TLSocketServerSentEvent } from '../protocol'
+import { TLPersistentClientSocketStatus } from '../TLSyncClient'
+
+/**
+ * Subset of `BroadcastChannel` used by the cross-tab machinery. Defined as
+ * an interface so tests can inject a mock that gossips between in-process
+ * instances.
+ *
+ * @internal
+ */
+export interface BroadcastChannelLike {
+	postMessage(msg: any): void
+	close(): void
+	addEventListener(type: 'message', handler: (ev: MessageEvent) => void): void
+	removeEventListener(type: 'message', handler: (ev: MessageEvent) => void): void
+}
+
+/**
+ * Subset of `navigator.locks` used by the cross-tab machinery.
+ *
+ * @internal
+ */
+export interface CrossTabLockManager {
+	request(
+		name: string,
+		options: { mode: 'exclusive'; signal?: AbortSignal },
+		callback: () => Promise<unknown>
+	): Promise<unknown>
+}
+
+/**
+ * Internal channel message envelope. The `_ct` tag namespaces our messages so
+ * a future user of the same channel doesn't accidentally trip our handlers.
+ *
+ * @internal
+ */
+export type CrossTabMessage =
+	| { _ct: 'leader-status'; status: TLPersistentClientSocketStatus; reason?: string }
+	| { _ct: 'leader-hello' }
+	| { _ct: 'follower-hello'; fromTabId: string }
+	| { _ct: 'client'; fromTabId: string; msg: TLSocketClientSentEvent<TLRecord> }
+	| { _ct: 'server-all'; msg: TLSocketServerSentEvent<TLRecord> }
+	| { _ct: 'server-to'; toTabId: string; msg: TLSocketServerSentEvent<TLRecord> }
+	| {
+			_ct: 'server-broadcast-except'
+			exceptTabId: string
+			msg: TLSocketServerSentEvent<TLRecord>
+	  }
+	/**
+	 * Sent on `focus` events. The pair `(claim, tabId)` orders all presenter
+	 * claims in the pool — the highest tuple wins. Claims use a monotonic
+	 * counter bumped past the highest seen value, so out-of-order delivery
+	 * over the channel doesn't matter.
+	 */
+	| { _ct: 'presenter-claim'; tabId: string; claim: number }
+	/**
+	 * Sent on `visibilitychange` (and once on construction). Lets each tab
+	 * maintain a set of currently-visible peers so the leader knows whether
+	 * to release its lock when it becomes hidden.
+	 */
+	| { _ct: 'visibility'; tabId: string; visible: boolean }
+	/**
+	 * Asks all peers to broadcast their current visibility, so a tab joining
+	 * mid-session can populate its peer-visibility set without waiting for
+	 * the next `visibilitychange`.
+	 */
+	| { _ct: 'visibility-request'; fromTabId: string }
+
+/**
+ * Thin typed wrapper over {@link BroadcastChannelLike}. Modules subscribe
+ * independently and filter by message type, instead of one giant switch.
+ *
+ * @internal
+ */
+export interface CrossTabChannel {
+	send(msg: CrossTabMessage): void
+	subscribe(handler: (msg: CrossTabMessage) => void): () => void
+	close(): void
+}
+
+/**
+ * Wrap a raw {@link BroadcastChannelLike} as a {@link CrossTabChannel}. Drops
+ * messages that don't match our envelope so a future co-tenant on the same
+ * channel name doesn't accidentally trip our handlers.
+ *
+ * @internal
+ */
+export function createCrossTabChannel(bc: BroadcastChannelLike): CrossTabChannel {
+	const handlers = new Set<(msg: CrossTabMessage) => void>()
+	const onMessage = (ev: MessageEvent) => {
+		const data = ev.data
+		if (!data || typeof data !== 'object' || !('_ct' in data)) return
+		handlers.forEach((h) => h(data as CrossTabMessage))
+	}
+	bc.addEventListener('message', onMessage)
+	return {
+		send(msg) {
+			bc.postMessage(msg)
+		},
+		subscribe(handler) {
+			handlers.add(handler)
+			return () => {
+				handlers.delete(handler)
+			}
+		},
+		close() {
+			bc.removeEventListener('message', onMessage)
+			bc.close()
+			handlers.clear()
+		},
+	}
+}
